09_tabular: Add note about bug when splitting data
As reported in #325, there's a mistake in the predicate that splits the data into the training and validation sets. Jeremy commented that it won't be fixed in this edition: https://github.com/fastai/fastbook/pull/337#issuecomment-735401046

This PR adds an errata comment to the notebook so that readers are aware of the mistake. This can save them time if they run into the bug on their own.